requirements.txt
djangorestframework==3.15.2
django-cors-headers==4.4.0
dj-database-url==2.1.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
argon2-cffi==23.1.0
drf-yasg==1.21.7
django-filter==24.3
python-decouple==3.8
reportlab==4.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
black==24.4.2
flake8==7.0.0
isort==5.13.2
pytest==8.1.1
pytest-django==4.8.0
Faker==24.8.0
# psycopg2-binary==2.9.9